Streamix notebook #streamix
Connection only to COM Port 3
Paste link to any youtube live stream that is currently 'live' and press connect button. All that on the condition that your arduino is already connected to USB. Than on your YouTube live stream send commands through comments - in that case numbers from 0-20 are all currently available commands that change the colours of neopixel rings.